Move functions for building components to their own file
